Use posthog in our docs
The idea is to use posthog for analytics of our docs, but to for it to be sent to the same posthog project as the rest of snaplet production (web, cli, www). In order to tell that the events came from the docs, I added `source: 'docs'` as a [super property](https://posthog.com/docs/libraries/js#super-properties).

I initially tried using the official [posthog-docusaurus integration plugin](https://github.com/PostHog/posthog-docusaurus). Problem is, this didn't give us a way to set up that `source: 'docs'` super property, so I copied the source, and modified it a little so that we can do this. We can fork the plugin and send the change upstream if we think its worth doing though.